// src/interfaces/Product.interface.ts
export interface Product {
  Id: string
  Denumire: string
  Pret_cu_TVA: string
  key?: string
  count?: number
  Alt_Cod?: string
  Categorie?: string
  Cod_EAN?: string
  Cod_SKU?: string
  Cota_TVA?: string
  Moneda?: string
  Observatii1?: string
  Observatii2?: string
  Observatii3?: string
  Pretul_fara_TVA?: string
  Tip?: string
  UM?: string
  prod_greutate?: string
}